Study Buddy-MCP/pyproject.toml
[project]
name = "studybuddy-mcp"
version = "0.1.0"
description = "AI-powered study companion with MCP and Gemini integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "gradio>=5.0.0",
    "google-generativeai>=0.8.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
